refractored commands.py to introudce plugins
